Create Taxi-v3_Q_net
This is an example of the classic Q network training and playing in openAI's gym environment. The game is Taxi_v3 (https://gym.openai.com/envs/Taxi-v3/), which is a game that is played on the terminal. As can be seen in the code, the agent first learns to pick up and drop off a passenger for the first time after 200 episodes. If you want many examples of pickup and drop off you will need to increase the number of episodes. I like to use 500 episodes. 

Training is done really quickly with this Q network and takes up a small amount of ram because it is a very simple game with a small amount of possible states. I believe it should be able to run, even if you have a small amount of ram available. 

Hope you enjoy it and you can use it to learn more about reinforcement learning. I know I did!