Update DQN
This code has been confirmed to work. I trained it for 1500 episodes, takes about 3 hours to train 500 episodes. I proved to myself it worked by appending the positive rewards to a list and printing them after the training has completed (EpiX), which positive rewards only happen after the agent successfully picks up and drops off a passenger winning the game.  The agent isn't perfect after 1500 episodes though, so if you want to take this code and improve it, by all means do so! Experiment and improve this code as you wish :).